Numerous changes to accommodate mcu version, see full text for details.
Restructure the crate to be a library crate with examples. The original
main is now rolled into an example with SDL2 as the backend. Also, add
sdl2 as a development dependency - this allows the library crate to not
depend on sdl2.

Remove dependency on rand crate - there was an issue compiling this on
the mcu version. This will likely become a parameter to Game::new().
For now, the piece order will be deterministic.

Make changes to the drawing method since the LCD on the boosterpack is
slow to refresh. This changes are incomplete. The new changes paint
over the old position of the piece before painting the current location.
Nothing is done for the settled spots or for when lines are cleared yet.